#14 Mapping Liste erstellt für vorgefertigte Icons
