+++ app/src/main/java/com/example/androidlauncher/ui/Utils.kt
@@ -38,6 +38,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.TransformOrigin
 import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.platform.LocalDensity
@@ -45,6 +46,7 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
+import com.composables.icons.lucide.Lucide
 import com.example.androidlauncher.data.AppInfo
 import com.example.androidlauncher.data.IconManager
 import com.example.androidlauncher.ui.theme.LocalDarkTextEnabled
@@ -52,6 +54,25 @@ import com.example.androidlauncher.ui.theme.LocalIconSize
 import kotlin.math.max
 import kotlin.math.roundToInt
 
+/**
+ * Default system-side mapping for app icons to Lucide icons.
+ * packageName -> lucideIconName
+ */
+val DEFAULT_ICON_MAPPINGS = mapOf(
+    "com.android.chrome" to "Globe",
+    "com.google.android.apps.messaging" to "MessageSquare",
+    "com.google.android.gm" to "Mail",
+    "com.google.android.calendar" to "Calendar",
+    "com.google.android.apps.photos" to "Image",
+    "com.android.settings" to "Settings",
+    "com.google.android.calculator" to "Calculator",
+    "com.google.android.apps.maps" to "MapPin",
+    "com.android.vending" to "ShoppingBag", // Play Store
+    "com.google.android.youtube" to "Youtube",
+    "com.whatsapp" to "Phone",
+    "com.instagram.android" to "Instagram",
+    "com.facebook.katana" to "Facebook"
+)
 
 /**
  * Extension function to find the Activity context from a Context.
@@ -98,31 +119,64 @@ fun AppIconView(app: AppInfo, modifier: Modifier = Modifier) {
     val isDarkTextEnabled = LocalDarkTextEnabled.current
     val tintColor = if (isDarkTextEnabled) Color.Black else Color.White
 
-    val customIconName = customIcons[app.packageName]
-    val customLucideIcon = if (customIconName != null) getLucideIconByName(customIconName) else null
+    // Priority: 1. User choice, 2. System default mapping, 3. App's own lucideIcon (if any)
+    val customIconName = customIcons[app.packageName] ?: DEFAULT_ICON_MAPPINGS[app.packageName]
+    val lucideIcon = if (customIconName != null) getLucideIconByName(customIconName) else app.lucideIcon
 
     Box(
         modifier = modifier.size(iconSize),
         contentAlignment = Alignment.Center
     ) {
         when {
-            customLucideIcon != null -> {
-                // Skalierung der Lucide Icons auf ca. 60% der Originalgröße.
+            lucideIcon != null -> {
+                // Skalierung der Lucide Icons auf ca. 55% der Originalgröße.
                 // Wir halten den Container (Box) auf iconSize und zentrieren das Icon darin,
                 // damit die Positionierung symmetrisch zu den anderen Icons bleibt.
                 Icon(
-                    imageVector = customLucideIcon,
+                    imageVector = lucideIcon,
                     contentDescription = null,
                     modifier = Modifier.size(iconSize * 0.55f),
                     tint = tintColor
                 )
             }
-            app.lucideIcon != null -> Icon(imageVector = app.lucideIcon, contentDescription = null, modifier = Modifier.size(iconSize), tint = tintColor)
             app.customIconResId != null -> Icon(painter = painterResource(id = app.customIconResId), contentDescription = null, modifier = Modifier.size(iconSize), tint = tintColor)
             app.iconBitmap != null -> Image(bitmap = app.iconBitmap, contentDescription = null, modifier = Modifier.size(iconSize), colorFilter = ColorFilter.tint(tintColor))
             else -> Box(modifier = Modifier.size(iconSize).background(tintColor.copy(alpha = 0.05f), CircleShape))
         }
     }
+}
+
+/**
+ * Retrieves a Lucide icon by name.
+ * Handles both direct members and extension properties (which are compiled to <Name>Kt classes).
+ */
+fun getLucideIconByName(name: String): ImageVector? {
+    val lucideClass = Lucide::class.java
+    
+    // 1. Try as a direct static member (Field or Method)
+    try {
+        val field = lucideClass.getField(name)
+        return field.get(null) as? ImageVector
+    } catch (e: Exception) {}
+    
+    try {
+        val methodName = if (name.startsWith("get")) name else "get$name"
+        val method = lucideClass.getMethod(methodName)
+        return method.invoke(null) as? ImageVector
+    } catch (e: Exception) {}
+
+    // 2. Try as a Kotlin extension property (compiled to com.composables.icons.lucide.<Name>Kt)
+    try {
+        // The naming convention for extension properties is usually IconNameKt
+        val className = "com.composables.icons.lucide.${name}Kt"
+        val clazz = Class.forName(className)
+        // Extension property "val Lucide.IconName" becomes "public static final ImageVector getIconName(Lucide receiver)"
+        val getterName = "get$name"
+        val method = clazz.getMethod(getterName, Lucide::class.java)
+        return method.invoke(null, Lucide) as? ImageVector
+    } catch (e: Exception) {}
+
+    return null
 }
 
 /**
